chore(suite-desktop): move discovery to common
