+++ packages/suite/src/actions/suite/__tests__/storageActions.test.ts
@@ -1,10 +1,15 @@
 import { Middleware } from 'redux';
 
-import { disableAccountsThunk, transactionsActions } from '@suite-common/wallet-core';
+import * as discoveryActions from '@suite-common/wallet-core';
+import {
+    prepareDiscoveryReducer,
+    disableAccountsThunk,
+    transactionsActions,
+    createDiscoveryThunk,
+} from '@suite-common/wallet-core';
 import { getAccountTransactions, getAccountIdentifier } from '@suite-common/wallet-utils';
 
 import * as walletSettingsActions from 'src/actions/settings/walletSettingsActions';
-import * as discoveryActions from 'src/actions/wallet/discoveryActions';
 import * as SUITE from 'src/actions/suite/constants/suiteConstants';
 import { accountsReducer, fiatRatesReducer, transactionsReducer } from 'src/reducers/wallet';
 import walletSettingsReducer from 'src/reducers/wallet/settingsReducer';
@@ -18,9 +23,7 @@ import { configureStore } from 'src/support/tests/configureStore';
 import { AppState } from 'src/types/suite';
 import { SETTINGS } from 'src/config/suite';
 import { preloadStore } from 'src/support/suite/preloadStore';
-import { prepareDiscoveryReducer } from 'src/reducers/wallet/discoveryReducer';
 import { extraDependencies } from 'src/support/extraDependencies';
-import { createDiscoveryThunk } from 'src/actions/wallet/discoveryThunks';
 
 import * as suiteActions from '../suiteActions';
 import * as storageActions from '../storageActions';
