+++ packages/suite/src/actions/wallet/selectedAccountActions.ts
@@ -1,16 +1,14 @@
+import { accountsActions, blockchainActions, discoveryActions } from '@suite-common/wallet-core';
+import { getAccountNetwork } from '@suite-common/wallet-utils';
+import { SelectedAccountStatus } from '@suite-common/wallet-types';
+import { DiscoveryStatus } from '@suite-common/wallet-constants';
+
 import { ROUTER, SUITE } from 'src/actions/suite/constants';
 import { selectDiscoveryForDevice } from 'src/reducers/suite/suiteReducer';
 import * as metadataActions from 'src/actions/suite/metadataActions';
 import * as comparisonUtils from 'src/utils/suite/comparisonUtils';
 import { getSelectedAccount } from 'src/utils/wallet/accountUtils';
 import { Action, Dispatch, GetState, AppState } from 'src/types/suite';
-
-import { accountsActions, blockchainActions } from '@suite-common/wallet-core';
-import { getAccountNetwork } from '@suite-common/wallet-utils';
-import { SelectedAccountStatus } from '@suite-common/wallet-types';
-import { DiscoveryStatus } from '@suite-common/wallet-constants';
-
-import { discoveryActions } from './discoveryActions';
 
 const getAccountState = (state: AppState): SelectedAccountStatus => {
     const { device } = state.suite;
