+++ packages/suite/src/components/wallet/AccountException/DiscoveryFailed.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 
+import { restartDiscoveryThunk } from '@suite-common/wallet-core';
+
 import { Translation } from 'src/components/suite';
 import { AccountExceptionLayout } from 'src/components/wallet';
 import { useDiscovery, useDispatch } from 'src/hooks/suite';
-import { restartDiscoveryThunk } from 'src/actions/wallet/discoveryThunks';
 
 /**
  * Handler for discovery "hard" error (other than bundle-error)
