+++ packages/suite/src/hooks/suite/__tests__/useDiscovery.test.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
-import { act } from '@testing-library/react';
-import { renderWithProviders } from 'src/support/tests/hooksHelper';
 
+import { act } from '@testing-library/react';
+import { combineReducers } from 'redux';
+
+import { prepareDiscoveryReducer } from '@suite-common/wallet-core';
+import { configureMockStore } from '@suite-common/test-utils';
+
+import { renderWithProviders } from 'src/support/tests/hooksHelper';
 import suiteReducer from 'src/reducers/suite/suiteReducer';
+import { extraDependencies } from 'src/support/extraDependencies';
+
 import { useDiscovery } from '../useDiscovery';
 import { actions } from '../__fixtures__/useDiscovery';
-import { configureMockStore } from '@suite-common/test-utils';
-import { combineReducers } from 'redux';
-import { prepareDiscoveryReducer } from 'src/reducers/wallet/discoveryReducer';
-import { extraDependencies } from 'src/support/extraDependencies';
 
 const discoveryReducer = prepareDiscoveryReducer(extraDependencies);
