+++ packages/suite/src/hooks/suite/useDiscovery.ts
@@ -1,10 +1,10 @@
 import { useCallback } from 'react';
 
+import { selectDiscoveryByDeviceState } from '@suite-common/wallet-core';
+import { DiscoveryStatus } from '@suite-common/wallet-constants';
+
 import { DiscoveryStatusType } from 'src/types/wallet';
 import { selectDevice } from 'src/reducers/suite/suiteReducer';
-import { selectDiscoveryByDeviceState } from 'src/reducers/wallet/discoveryReducer';
-
-import { DiscoveryStatus } from '@suite-common/wallet-constants';
 
 import { useSelector } from './useSelector';
