+++ packages/suite/src/middlewares/suite/logsMiddleware.ts
@@ -1,4 +1,11 @@
 import { MiddlewareAPI } from 'redux';
+
+import { discoveryActions } from '@suite-common/wallet-core';
+import { addLog } from '@suite-common/logger';
+import { TRANSPORT, DEVICE } from '@trezor/connect';
+import { redactUserPathFromString } from '@trezor/utils';
+import { analyticsActions } from '@suite-common/analytics';
+
 import { AppState, Action, Dispatch } from 'src/types/suite';
 import {
     DESKTOP_UPDATE,
@@ -11,12 +18,6 @@ import {
 import { WALLET_SETTINGS } from 'src/actions/settings/constants';
 import * as walletSettingsActions from 'src/actions/settings/walletSettingsActions';
 import { redactTransactionIdFromAnchor } from 'src/utils/suite/analytics';
-import { discoveryActions } from 'src/actions/wallet/discoveryActions';
-
-import { addLog } from '@suite-common/logger';
-import { TRANSPORT, DEVICE } from '@trezor/connect';
-import { redactUserPathFromString } from '@trezor/utils';
-import { analyticsActions } from '@suite-common/analytics';
 
 const log =
     (api: MiddlewareAPI<Dispatch, AppState>) =>
