+++ packages/suite/src/middlewares/wallet/graphMiddleware.ts
@@ -1,11 +1,11 @@
 import { MiddlewareAPI } from 'redux';
+
+import { discoveryActions, accountsActions, transactionsActions } from '@suite-common/wallet-core';
+import { DiscoveryStatus } from '@suite-common/wallet-constants';
+
 import * as graphActions from 'src/actions/wallet/graphActions';
 import { selectDiscoveryForDevice } from 'src/reducers/suite/suiteReducer';
 import { AppState, Action, Dispatch } from 'src/types/suite';
-import { discoveryActions } from 'src/actions/wallet/discoveryActions';
-
-import { accountsActions, transactionsActions } from '@suite-common/wallet-core';
-import { DiscoveryStatus } from '@suite-common/wallet-constants';
 
 const graphMiddleware =
     (api: MiddlewareAPI<Dispatch, AppState>) =>
