+++ suite-common/wallet-core/tsconfig.json
@@ -10,6 +10,7 @@
         { "path": "../test-utils" },
         { "path": "../toast-notifications" },
         { "path": "../wallet-config" },
+        { "path": "../wallet-constants" },
         { "path": "../wallet-types" },
         { "path": "../wallet-utils" },
         {
@@ -19,6 +20,7 @@
             "path": "../../packages/blockchain-link-utils"
         },
         { "path": "../../packages/connect" },
+        { "path": "../../packages/device-utils" },
         { "path": "../../packages/type-utils" },
         { "path": "../../packages/utils" }
     ]
